move AssuranceOptions to assurance class file move flatten_xml_to_dict, console_welcome_banner, find_close_matches, resolve_hostname to utilities
